+++ src/App.jsx
@@ -115,6 +115,7 @@ const MisCreacionesPage = lazy(() => import('./pages/cuenta/MisCreacionesPage'))
 const CuentaReferidosPage = lazy(() => import('./pages/cuenta/CuentaReferidosPage'));
 const CuentaFechasImportantesPage = lazy(() => import('./pages/cuenta/CuentaFechasImportantesPage'));
 const CatalogReward = lazy(() => import('./pages/cuenta/CatalogReward'));
+const MisionesPage = lazy(() => import('./pages/cuenta/MisionesPage'));
 const WishlistPrivatePage = lazy(() => import('./pages/cuenta/WishlistPage'));
 const WishlistPublicPage = lazy(() => import('./pages/WishlistPublic/WishlistPublic'));
 
@@ -226,6 +227,7 @@ function App() {
                           <Route path="creaciones" element={<MisCreacionesPage />} />
                           <Route path="referidos" element={<CuentaReferidosPage />} />
                           <Route path="fechas-importantes" element={<CuentaFechasImportantesPage />} />
+                          <Route path="misiones" element={<MisionesPage />} />
                           <Route path="catalogo" element={<CatalogReward />} />
                           <Route path="wishlist" element={<WishlistPrivatePage />} />
                         </Route>